refactored modal to take children
