+++ src/components/EventCardCollection.tsx
@@ -1,6 +1,6 @@
 // React Imports
 import React, { useState } from 'react';
-import { useQuery } from '@apollo/react-hooks';
+import { useQuery, useMutation } from '@apollo/react-hooks';
 import gql from 'graphql-tag';
 
 // Custom Imports
@@ -9,6 +9,7 @@ import notFoundGif from '../media/animations/event-not-found.gif';
 import Modal from '../components/Modal';
 //file generated by running npm run apollo:generate, this looks at our graphql schema and creates types for typescript based on the schema
 import { getEvents } from '../generated/getEvents';
+import { getEvents_getEvents } from '../generated/getEvents';
 // This component holds multiple Event Cards accessible through scrolling
 // Layout:
 // List of Events
@@ -30,15 +31,39 @@ const GET_EVENTS = gql`
   }
 `;
 
+//graphql mutation to sign up volunteers for an event
+const ADD_VOLUNTEER = gql`
+  mutation addVolunteer($volunteerId: Int!, $eventId: Int!) {
+    addVolunteer(volunteerId: $volunteerId, eventId: $eventId) {
+      ok
+      errors {
+        path
+        message
+      }
+    }
+  }
+`;
+
 const EventCardCollection = () => {
   //useQuery hook takes our gql query and makes a request to the backend
   //the return value is held in data
   const [isModalOpen, setModalOpen] = useState(false);
-  const [modalContent, setModalContent] = useState({});
+  const [selectedEventId, setSelectedEventId] = useState<number | null>(null);
+  const [addVolunteer] = useMutation(ADD_VOLUNTEER);
 
-  const { loading, error, data } = useQuery<getEvents>(GET_EVENTS);
-
-  if (loading || !data) return null;
+  const handleAttendClick = async () => {
+    const volunteerId = parseInt(localStorage.getItem('volunteerId')!);
+    let mutationResult = await addVolunteer({
+      variables: {
+        eventId: selectedEventId,
+        volunteerId,
+      },
+    });
+    if (mutationResult.data.addVolunteer.ok) {
+      closeModal();
+      //add notfication of successful signup
+    }
+  };
 
   const toggleModal = () => {
     setModalOpen(!isModalOpen);
@@ -48,19 +73,37 @@ const EventCardCollection = () => {
     setModalOpen(false);
   };
 
-  const addModalContent = (id: number) => {
-    const selectedEventData = data.getEvents.filter(event => event.id === id);
-    setModalContent(selectedEventData[0]);
-    toggleModal();
-  };
+  const { loading, error, data } = useQuery<getEvents>(GET_EVENTS);
+
+  if (loading || !data) return null;
+  const { getEvents: events } = data;
+
+  const selectedEvent = !!selectedEventId
+    ? events.find(event => event.id === selectedEventId)
+    : null;
 
   return (
     <div>
-      <Modal
-        isModalOpen={isModalOpen}
-        modalContent={modalContent}
-        closeModal={closeModal}
-      />
+      <Modal isModalOpen={isModalOpen} closeModal={closeModal}>
+        <p>
+          {selectedEvent?.title} by {selectedEvent?.nonprofit['displayName']}
+        </p>
+        <p>{selectedEvent?.date}</p>
+        <p>{selectedEvent?.location}</p>
+        <button
+          onClick={handleAttendClick}
+          style={{
+            position: 'absolute',
+            bottom: 5,
+            right: 10,
+            color: 'white',
+            background: '#2a4365',
+            padding: '5px',
+            borderRadius: '15px',
+          }}>
+          Attend
+        </button>
+      </Modal>
       {!error ? (
         <div>
           <div>
@@ -70,7 +113,10 @@ const EventCardCollection = () => {
                 <EventCard
                   event={event}
                   key={event.id}
-                  addModalContent={addModalContent}
+                  addModalContent={() => {
+                    setSelectedEventId(event.id);
+                  }}
+                  toggleModal={toggleModal}
                 />
               );
             })}
